+++ src/main/java/moze_intel/projecte/gameObjs/items/ItemPE.java
@@ -20,34 +20,34 @@ public abstract class ItemPE extends Item
 		return super.setUnlocalizedName("pe_" + message);
 	}
 
-	public static double getEmc(ItemStack stack)
+	public static long getEmc(ItemStack stack)
 	{
 		if (stack.stackTagCompound == null)
 		{
 			stack.stackTagCompound = new NBTTagCompound();
 		}
 		
-		return stack.stackTagCompound.getDouble("StoredEMC");
+		return stack.stackTagCompound.getLong("StoredEMC");
 	}
 	
-	public static void setEmc(ItemStack stack, double amount)
+	public static void setEmc(ItemStack stack, long amount)
 	{
 		if (stack.stackTagCompound == null)
 		{
 			stack.stackTagCompound = new NBTTagCompound();
 		}
 		
-		stack.stackTagCompound.setDouble("StoredEMC", amount);
+		stack.stackTagCompound.setLong("StoredEMC", amount);
 	}
 	
-	public static void addEmcToStack(ItemStack stack, double amount)
+	public static void addEmcToStack(ItemStack stack, long amount)
 	{
 		setEmc(stack, getEmc(stack) + amount);
 	}
 	
-	public static void removeEmc(ItemStack stack, double amount)
+	public static void removeEmc(ItemStack stack, long amount)
 	{
-		double result = getEmc(stack) - amount;
+		long result = getEmc(stack) - amount;
 		
 		if (result < 0)
 		{
@@ -57,18 +57,18 @@ public abstract class ItemPE extends Item
 		setEmc(stack, result);
 	}
 	
-	public static boolean consumeFuel(EntityPlayer player, ItemStack stack, double amount, boolean shouldRemove)
+	public static boolean consumeFuel(EntityPlayer player, ItemStack stack, long amount, boolean shouldRemove)
 	{
 		if (amount <= 0)
 		{
 			return true;
 		}
 
-		double current = getEmc(stack);
+		long current = getEmc(stack);
 		
 		if (current < amount)
 		{
-			double consume = EMCHelper.consumePlayerFuel(player, amount - current);
+			long consume = EMCHelper.consumePlayerFuel(player, amount - current);
 			
 			if (consume == -1)
 			{
